refactor(View): 출력문 enum 클래스 리팩터링
view 에서만 사용되는 출력문을 관리하는 enum 클래스를 view 디렉터리로 옮긴다.

변경사항:
출력 메소드명 printMessage -> print
