+++ src/main/java/bridge/view/InputView.java
@@ -1,35 +1,23 @@
 package bridge.view;
 
-import bridge.domain.stringenum.PrintMessage;
+import bridge.view.stringenum.PrintMessage;
 import bridge.util.InputValidator;
 import camp.nextstep.edu.missionutils.Console;
 
-/**
- * 사용자로부터 입력을 받는 역할을 한다.
- */
 public class InputView {
 
-    /**
-     * 다리의 길이를 입력받는다.
-     */
     public int readBridgeSize() {
-        printMessage(PrintMessage.BRIDGE_LENGTH.printMessage());
+        printMessage(PrintMessage.BRIDGE_LENGTH.print());
         return readNumber(Console.readLine());
     }
 
-    /**
-     * 사용자가 이동할 칸을 입력받는다.
-     */
     public String readMoving() {
-        printMessage(PrintMessage.MOVE.printMessage());
+        printMessage(PrintMessage.MOVE.print());
         return readString(Console.readLine());
     }
 
-    /**
-     * 사용자가 게임을 다시 시도할지 종료할지 여부를 입력받는다.
-     */
     public String readGameCommand() {
-        printMessage(PrintMessage.RESTART.printMessage());
+        printMessage(PrintMessage.RESTART.print());
         return readString(Console.readLine());
     }
