+++ src/main/java/bridge/view/OutputView.java
@@ -1,37 +1,24 @@
 package bridge.view;
 
-import bridge.domain.stringenum.PrintMessage;
+import bridge.view.stringenum.PrintMessage;
 import bridge.util.MapConverter;
 import java.util.List;
 
-/**
- * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
- */
 public class OutputView {
 
-    /**
-     * 현재까지 이동한 다리의 상태를 정해진 형식에 맞춰 출력한다.
-     * <p>
-     * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
-     */
-    public void printMap(List<List<String>> crossedBridge) {
-        crossedBridge.forEach(this::printEachMap);
+    public void printMap(List<List<String>> moveResult) {
+        moveResult.forEach(this::printEachMap);
         System.out.println();
     }
 
-    /**
-     * 게임의 최종 결과를 정해진 형식에 맞춰 출력한다.
-     * <p>
-     * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
-     */
-    public void printResult(List<List<String>> crossedBridge, int tryCount, boolean isSuccess) {
-        System.out.println(PrintMessage.END.printMessage());
-        printMap(crossedBridge);
+    public void printResult(List<List<String>> moveResult, int tryCount, boolean isSuccess) {
+        System.out.println(PrintMessage.END.print());
+        printMap(moveResult);
         printFinalGameResult(tryCount, isSuccess);
     }
 
     public void printStart() {
-        System.out.println(PrintMessage.START.printMessage() + "\n");
+        System.out.println(PrintMessage.START.print() + "\n");
     }
 
     private void printEachMap(List<String> crossedBridge) {
@@ -39,7 +26,7 @@ public class OutputView {
     }
 
     private void printFinalGameResult(int tryCount, boolean isSuccess) {
-        System.out.println(PrintMessage.RESULT.printMessage(isSuccess));
-        System.out.println(PrintMessage.TRY.printMessage(tryCount));
+        System.out.println(PrintMessage.RESULT.print(isSuccess));
+        System.out.println(PrintMessage.TRY.print(tryCount));
     }
 }
